Start using ORT 1.20 kernels (#97)

+++ ndonnx/_core/_numericimpl.py
@@ -375,7 +375,7 @@ class _NumericOperationsImpl(OperationsBlock):
             ),
             [ndx.reshape(x, [-1]) if axis is None else x],
             cast_return=False,
-            int_dtype=ndx.int32,
+            int_dtype=ndx.int64,
             float_dtype=ndx.float64,
         )
 
@@ -393,7 +393,7 @@ class _NumericOperationsImpl(OperationsBlock):
             ),
             [ndx.reshape(x, [-1]) if axis is None else x],
             cast_return=False,
-            int_dtype=ndx.int32,
+            int_dtype=ndx.int64,
             float_dtype=ndx.float64,
         )
